# service/api/update-group-photo.go
package api

import (
	"bytes"
	"github.com/julienschmidt/httprouter"
	"io"
	"net/http"
	"os"
	// "strconv"
	// "strings"
	"path/filepath"
)

func (rt *_router) putNewGroupPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
	// aut := r.Header.Get("Authorization")
	// valori := strings.Split(aut, " ")
	// chat_id := valori[1]

	/*
		w.Header().Set("Content-Type", "application/json")
		data, err := io.ReadAll(r.Body)
		if err != nil {
			w.WriteHeader(http.StatusInternalServerError)
			// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
			return
		}
		r.Body = io.NopCloser(bytes.NewBuffer(data))
		id, err := rt.GetIdGroupPicture(chat_id)
		if err != nil {
			w.WriteHeader(http.StatusInternalServerError)
			// ctx.Logger.WithError(err).Error("session: can't create response json")
			return
		}
		if id == 0 {
			id, err = rt.CreateNewPhotoId(chat_id)
			if err != nil {
				w.WriteHeader(http.StatusInternalServerError)
				// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
				return
			}
			_, err = os.Stat("./media/group_picture/" + chat_id)
			if os.IsNotExist(err) {
				// creare cartella
				err = os.Mkdir("./media/group_picture/"+chat_id, os.ModeDir)
				if err != nil {
					w.WriteHeader(http.StatusInternalServerError)
					// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
					return
				}

			}
		}

		// creare file
		id_string := strconv.FormatInt(id, 36) // ???
		out, err := os.Create("./media/group_picture/" + chat_id + "/" + id_string)
		if err != nil {
			w.WriteHeader(http.StatusInternalServerError)
			// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
			return
		}
		_, err = io.Copy(out, r.Body)
		if err != nil {
			w.WriteHeader(http.StatusInternalServerError)
			// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
			return
		}
		out.Close()
		w.WriteHeader(http.StatusOK)

	*/

	w.Header().Set("Content-Type", "application/json")
	data, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
		return
	}
	r.Body = io.NopCloser(bytes.NewBuffer(data))
	path := filepath.Join("/tmp/media", ps.ByName("chat_id"), "1.jpg")
	_, err = os.Stat(path)
	if err == nil {
		os.Remove(path)
	}
	// creare file
	out, err := os.Create(path)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
		return
	}
	_, err = io.Copy(out, r.Body)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		// ctx.Logger.WithError(err).Error("Can't retrieve photo data")
		return
	}
	out.Close()
	w.WriteHeader(http.StatusOK) // risposta

}